% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc.predictions.R
\name{drm.prediction.fx}
\alias{drm.prediction.fx}
\title{Creates a predictions table to create 95% confidence intervals
  that can be used for plotting dose response curves in ggplot2.}
\usage{
drm.prediction.fx(model, doses, curvevar)
}
\arguments{
\item{model}{object. drm model}

\item{doses}{numerical sequence vector of the range of doses tested}

\item{curvevar}{character string of the variable curves are predicted for.}
}
\description{
Creates a predictions table to create 95% confidence intervals
  that can be used for plotting dose response curves in ggplot2.
}
