% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc.predictions.R
\name{drm.prediction}
\alias{drm.prediction}
\title{Creates a predictions table to create 95% confidence intervals
  that can be used for plotting dose response curves in ggplot2.}
\usage{
drm.prediction(model, length = 100)
}
\arguments{
\item{model}{object. drm model}

\item{length}{numeric. length of doses to form prediction table}
}
\description{
Creates a predictions table to create 95% confidence intervals
  that can be used for plotting dose response curves in ggplot2.
}
