% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germinator-package.R
\docType{package}
\name{Germinator}
\alias{Germinator}
\title{Germinator: a package for processing "HyphaTracker" data output.}
\description{
"HyphaTracker" is an ImageJ macro plugin developed by \cite{Brunk et al 2018}.
}
\details{
The Germinator package provides four important functions: Design of experiment
 and summary of data for plotting and model fitting.
}
\section{Germinator functions, Lookup.maker}{
 creates the plate well treatment
 assignments and outputs the lookup table of the treatments. Growth: Assigns
 variable factors to "HyphaTracker" data. Summarizes the average amount of
 growth for each well and then calculates growth.
}

