% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth.R
\name{Growth}
\alias{Growth}
\title{Process and summarize data from "HyphaTracker" using the design from
Plate.Design}
\usage{
Growth(data = NULL, Lookup = NULL, badimage = c(""))
}
\arguments{
\item{data}{Dataframe from "HyphaTracker" output.}

\item{Lookup}{Dataframe lookup table from the `Plate.Design` output}

\item{badimage}{logical, if their are images to be removed from analysis}
}
\value{
An object of class 'Germinator'
}
\description{
Process and summarize data from "HyphaTracker" using the design from
Plate.Design
}
