% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc.predictions.R
\name{drc.predictions}
\alias{drc.predictions}
\title{Creates a predicitons table to create 95% confidence intervals
  that can be used for plotting dose response curves in ggplot2.}
\usage{
drc.predictions(
  data,
  splitfct,
  nest.fct,
  concrange = c(0, 100),
  smoothness = 100,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{splitfct}{variable to split data by. as chaaracter}

\item{nest.fct}{variable to nest data by}

\item{concrange}{numeric vector of min and max dose}

\item{smoothness}{Smoothness of prediction line. Default is 100.}

\item{...}{arguments to pass into drm}
}
\description{
Creates a predicitons table to create 95% confidence intervals
  that can be used for plotting dose response curves in ggplot2.
}
