% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc.predictions.R
\name{drc.predictions}
\alias{drc.predictions}
\title{Creates a predicitons table to create 95% confidence intervals
  that can be used for plotting dose response curves in ggplot2.}
\usage{
drc.predictions(
  data,
  model,
  conc,
  LOG = TRUE,
  sm.value = 5e-05,
  smoothness = 100
)
}
\arguments{
\item{data}{dataframe used for the model}

\item{model}{model object from drm}

\item{conc}{numeric string. Concentrations or doses used in acsending order.}

\item{LOG}{Logical. True if concentration scale is logistic. Default is TRUE.}

\item{sm.value}{Numeric. Small value to add to concentrations of zero to
avoid log of 0. Defualt is 0.00005.}

\item{smoothness}{resolution of confidence intervals. 100 is default}
}
\description{
Creates a predicitons table to create 95% confidence intervals
  that can be used for plotting dose response curves in ggplot2.
}
