% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Growth.R
\name{Growth}
\alias{Growth}
\title{Process and summarise data from "HyphaTracker" using the design from
Plate.Design}
\usage{
Growth(data, Lookup)
}
\arguments{
\item{data}{Dataframe from "HyphaTracker" output.}

\item{Lookup}{Dataframe lookup table from the Plate.Design output}
}
\value{
An object of class 'Germinator'
}
\description{
Process and summarise data from "HyphaTracker" using the design from
Plate.Design
}
