% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lookup.maker.R
\name{Lookuptable.maker}
\alias{Lookuptable.maker}
\title{Creating lookup table from vectors of fungicides and concentrations}
\usage{
Lookuptable.maker(
  Treatments = NULL,
  Isolates = NULL,
  Wells = 96,
  Reps = 4,
  timepts = 2,
  random = FALSE,
  randseed = 123456,
  snake = TRUE
)
}
\arguments{
\item{Treatments}{A vector of treatments. Pesiticide name and concentration
seperated by a single space.}

\item{Isolates}{A vector of isolates names.
In order in which they are imaged}

\item{Wells}{Numeric, how many wells on a plate. Default is 96}

\item{Reps}{Numeric, How many times is each treatment repeated on the plate.
Default is 4}

\item{timepts}{Numeric, timepoints evaluated. Default is 2.}

\item{random}{Logical, treatments randomly assigned to wells.
Default in FALSE}

\item{randseed}{Numeric, randomization seed to be used. Default is 123456.}

\item{snake}{Logical. Indicate if the plate was imaged in a snake
like pattern. Default is TRUE.}
}
\description{
Creating lookup table from vectors of fungicides and concentrations
}
